+++ GroamWorld/rotation.h
@@ -5,11 +5,12 @@ class Rotation
 	private:
 		int xPos;
 		int yPos;
-		float xRot;
-		float yRot;
+		int xRot;
+		int yRot;
+
 	public:
 		void init(int, int);
 		void rotate(int, int);
-		float getXRot();
-		float getYRot();
+		int getXRot();
+		int getYRot();
 };
